[********]<feature>核间负载数据解析功能头文件定义-修改检视意见
Match-id-87b538df2654fad4b646949b40fef3c3572e4c17

+++ server/src/utils/JsonUtil.h
@@ -188,7 +188,7 @@ public:
         }
         if (json[key.data()].IsString()) {
             try {
-                return std::stod(json[key.data()].GetString());
+                return std::stof(json[key.data()].GetString());
             } catch (std::exception &e) {
                 return 0;
             }
